camoufox-py/requirements.txt
PyYAML==6.0.2
camoufox[geoip]==0.4.11
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
browserforge==1.2.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.3
idna==3.10
language-tags==1.2.0
lxml==5.4.0
maxminddb==2.7.0
multidict==6.5.0
numpy==2.3.0
orjson==3.10.18
platformdirs==4.3.8
playwright==1.52.0
propcache==0.3.2
pyee==13.0.0
PySocks==1.7.1
requests==2.32.4
screeninfo==0.8.1
tqdm==4.67.1
typing_extensions==4.14.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
yarl==1.20.1